refactor(gameplay): dedupe FindActorById into PIESequenceFormat
The same actor lookup (by name / class name / full path / path suffix)
was copy-pasted into PIEInputRecorder, PIEInputReplayer, and
GameplayHandlers_PIESnapshot. Moves the canonical implementation into
PIESequenceFormat alongside the other shared on-disk helpers, drops the
three local copies, and the snapshot caller qualifies the namespace
since its using-directive is scoped to the file's anonymous namespace.
